Remove old New page
